Refactored Integration Tests (#53)
* Refactored Integration Tests

* Fixed linting errors

* Fixed linting errors again

* Fixed linting errors again again

* Fixed linting errors again again again

* linting is back

* linting is back again

* linting is back again again

* linting is back again again again

* linting is back again again again again
