+++ tests/simulation/sim_queries.go
@@ -1,6 +1,201 @@
+//nolint:all
 package simulation
 
-import "fmt"
+import (
+	"fmt"
+	"net/url"
+	"testing"
+	"time"
+
+	"github.com/jmoiron/sqlx"
+	_ "github.com/lib/pq"
+
+	"github.com/newrelic/infra-integrations-sdk/v3/log"
+	"github.com/stretchr/testify/require"
+)
+
+const (
+	defaultUser = "dbuser"
+	defaultPass = "dbpassword"
+	testDB      = "titanic"
+)
+
+func openDB(targetContainer string) (*sqlx.DB, error) {
+	dbPort := getPortForContainer(targetContainer)
+
+	connectionURL := &url.URL{
+		Scheme: "postgres",
+		User:   url.UserPassword(defaultUser, defaultPass),
+		Host:   fmt.Sprintf("%s:%d", "localhost", dbPort),
+		Path:   testDB,
+	}
+
+	query := url.Values{}
+	query.Add("connect_timeout", "10")
+	query.Add("sslmode", "disable")
+
+	connectionURL.RawQuery = query.Encode()
+	dsn := connectionURL.String()
+	db, err := sqlx.Open("postgres", dsn)
+	if err != nil {
+		return nil, fmt.Errorf("cannot connect to db: %s", err)
+	}
+	return db, nil
+}
+
+func getPortForContainer(container string) int {
+	switch container {
+	case "postgresql-perf-latest":
+		return 5432
+	case "postgresql-perf-oldest":
+		return 6432
+	case "postgresql-noext":
+		return 7432
+	default:
+		return 5432
+	}
+}
+
+// SimulationController handles coordinating multiple database simulations
+type SimulationController struct {
+	targetContainer string
+	envVars         []string
+}
+
+// NewSimulationController creates a new controller for database simulations
+func NewSimulationController(targetContainer string, envVars ...string) *SimulationController {
+	return &SimulationController{
+		targetContainer: targetContainer,
+		envVars:         envVars,
+	}
+}
+
+// StartAllSimulations starts all simulation routines concurrently
+func (sc *SimulationController) StartAllSimulations(t *testing.T) chan struct{} {
+	done := make(chan struct{})
+
+	go func() {
+		defer close(done)
+
+		// Create error channel to collect errors from goroutines
+		errChan := make(chan error, 6)
+
+		// Start all simulations in separate goroutines
+		go func() {
+			SimulateQueries(t, sc.targetContainer)
+			errChan <- nil
+		}()
+
+		go func() {
+			SimulateSlowQueries(t, sc.targetContainer)
+			errChan <- nil
+		}()
+
+		for pclass := 1; pclass <= 3; pclass++ {
+			go func() {
+				SimulateWaitEvents(t, sc.targetContainer, pclass)
+				errChan <- nil
+			}()
+		}
+
+		go func() {
+			SimulateBlockingSessions(t, sc.targetContainer)
+			errChan <- nil
+		}()
+
+		// Wait for all goroutines to complete
+		for i := 0; i < 6; i++ {
+			if err := <-errChan; err != nil {
+				log.Error("Error in simulation routine: %v", err)
+				t.Error(err)
+			}
+		}
+	}()
+
+	return done
+}
+
+func ExecuteQuery(t *testing.T, query string, targetContainer string, delay int) {
+	db, err := openDB(targetContainer)
+	require.NoError(t, err)
+	defer db.Close()
+
+	_, err = db.Exec(query)
+	require.NoError(t, err)
+	time.Sleep(time.Duration(delay) * time.Millisecond)
+}
+
+func SimulateQueries(t *testing.T, targetContainer string) {
+	t.Helper()
+	for _, query := range SimpleQueries() {
+		ExecuteQuery(t, query, targetContainer, 100)
+	}
+}
+
+func SimulateSlowQueries(t *testing.T, targetContainer string) {
+	t.Helper()
+	for _, query := range SlowQueries() {
+		ExecuteQuery(t, query, targetContainer, 500)
+	}
+}
+
+func SimulateWaitEvents(t *testing.T, targetContainer string, pclass int) {
+	t.Helper()
+
+	queries := WaitEventQueries(pclass)
+
+	// Start the locking transaction in a goroutine
+	go func() {
+		ExecuteQuery(t, queries.LockingQuery, targetContainer, 100)
+	}()
+
+	// Wait for first transaction started
+	time.Sleep(2 * time.Second)
+
+	// Run the blocked transaction
+	ExecuteQuery(t, queries.BlockedQuery, targetContainer, 100)
+
+	time.Sleep(30 * time.Second)
+}
+
+func SimulateBlockingSessions(t *testing.T, targetContainer string) {
+	t.Helper()
+
+	queries := BlockingQueries()
+
+	db, err := openDB(targetContainer)
+	require.NoError(t, err)
+	defer db.Close()
+
+	// Start the first transaction that will hold the lock
+	tx1, err := db.Begin()
+	require.NoError(t, err)
+	defer tx1.Rollback()
+
+	// Execute the locking query
+	_, err = tx1.Exec(queries.HoldLockQuery)
+	require.NoError(t, err)
+
+	// Start the blocking query in a separate goroutine
+	go func() {
+		time.Sleep(2 * time.Second) // Wait for a bit before trying to acquire lock
+
+		tx2, err := db.Begin()
+		if err != nil {
+			t.Error(err)
+			return
+		}
+		defer tx2.Rollback()
+
+		// This query will block waiting for tx1's lock
+		tx2.Exec(queries.BlockedQuery)
+		// We don't check for errors here since this might timeout
+	}()
+
+	// Hold the lock for a few seconds, then release it
+	time.Sleep(5 * time.Second)
+	tx1.Commit()
+}
 
 func SimpleQueries() []string {
 	return []string{
